Split a single score entry into a separate class for legibility and implemented drop-down functionality for each score update. Added a youtube embed to dropdown to test out its functionality
