+++ frontend/src/scenes/global/DropdownHandler.js
@@ -1,0 +1,11 @@
+import { useCallback, useState } from "react";
+
+export default function DropdownHandler(initialState) { 
+    const [isOpen, setOpenState] = useState(initialState)
+
+    const toggle = useCallback (() => {
+        setOpenState((state) => !state);
+    }, [setOpenState]);
+
+    return {isOpen, toggle}
+}
